Move capability_ptr to utilities
