+++ kernel/src/upcall.rs
@@ -4,12 +4,12 @@
 
 //! Data structure for storing an upcall from the kernel to a process.
 
-use crate::capability_ptr::CapabilityPtr;
 use crate::config;
 use crate::debug;
 use crate::process;
 use crate::process::ProcessId;
 use crate::syscall::SyscallReturn;
+use crate::utilities::capability_ptr::CapabilityPtr;
 use crate::ErrorCode;
 
 /// Type to uniquely identify an upcall subscription across all drivers.
