+++ kernel/src/utilities/mod.rs
@@ -14,6 +14,7 @@ pub mod peripheral_management;
 pub mod static_init;
 pub mod storage_volume;
 
+pub mod capability_ptr;
 mod static_ref;
 
 pub use self::static_ref::StaticRef;
